........S. [ZBX-13755] fixed proxy lastaccess update on 32 bit Zabbix server

+++ src/libs/zbxdbhigh/proxy.c
@@ -396,7 +396,7 @@ static void	db_update_proxies_lastaccess(const zbx_vector_uint64_pair_t *proxy_d
 		zbx_snprintf_alloc(&sql, &sql_alloc, &sql_offset, "update hosts"
 				" set lastaccess=%d"
 				" where hostid=" ZBX_FS_UI64 ";\n",
-				pair.second, pair.first);
+				(int)pair.second, pair.first);
 
 		DBexecute_overflowed_sql(&sql, &sql_alloc, &sql_offset);
 	}
